# docs/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User Transaction",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Service"
    }
  ],
  "host": "localhost:9000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/auth/users/{user_id}/transactions": {
      "get": {
        "operationId": "Service_ListTransactions",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionListTransactionsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "ID of user",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "description": "Bearer token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Service"
        ]
      },
      "post": {
        "operationId": "Service_CreateTransaction",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionCreateTransactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "ID of user",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ServiceCreateTransactionBody"
            }
          }
        ],
        "tags": [
          "Service"
        ]
      }
    },
    "/api/auth/users/{user_id}/transactions/{transaction_id}": {
      "delete": {
        "operationId": "Service_DeleteTransaction",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionDeleteTransactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "ID of user",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "transaction_id",
            "description": "ID of transaction",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Service"
        ]
      },
      "put": {
        "operationId": "Service_UpdateTransaction",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/transactionUpdateTransactionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user_id",
            "description": "ID of user",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ServiceUpdateTransactionBody"
            }
          }
        ],
        "tags": [
          "Service"
        ]
      }
    }
  },
  "definitions": {
    "ServiceCreateTransactionBody": {
      "type": "object",
      "properties": {
        "transaction": {
          "$ref": "#/definitions/transactionTransaction"
        }
      }
    },
    "ServiceUpdateTransactionBody": {
      "type": "object",
      "properties": {
        "transaction": {
          "$ref": "#/definitions/transactionTransaction"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "transactionCreateTransactionResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/transactionCreateTransactionResponseData"
        }
      }
    },
    "transactionCreateTransactionResponseData": {
      "type": "object",
      "properties": {
        "transaction": {
          "$ref": "#/definitions/transactionTransaction"
        }
      }
    },
    "transactionDecimal": {
      "type": "object",
      "properties": {
        "unit": {
          "type": "string",
          "format": "int64"
        },
        "nanos": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "transactionDeleteTransactionResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "transactionListTransactionsResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/transactionListTransactionsResponseData"
        }
      }
    },
    "transactionListTransactionsResponseData": {
      "type": "object",
      "properties": {
        "transactions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/transactionTransaction"
          }
        }
      }
    },
    "transactionTransaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "account_id": {
          "type": "string",
          "format": "int64"
        },
        "amount": {
          "$ref": "#/definitions/transactionDecimal"
        },
        "transaction_type": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "transactionUpdateTransactionResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/transactionUpdateTransactionResponseData"
        }
      }
    },
    "transactionUpdateTransactionResponseData": {
      "type": "object",
      "properties": {
        "transaction": {
          "$ref": "#/definitions/transactionTransaction"
        }
      }
    }
  }
}